+++ components/Icon.vue
@@ -1,9 +1,21 @@
 <script setup lang="ts">
+import circleQuestion from '../assets/icons/circle-question.svg?raw';
 import clone from '../assets/icons/clone.svg?raw';
+import ellipsisVertical from '../assets/icons/ellipsis-vertical.svg?raw';
+import fileExport from '../assets/icons/file-export.svg?raw';
+import fileImport from '../assets/icons/file-import.svg?raw';
 import trash from '../assets/icons/trash.svg?raw';
 import xmark from '../assets/icons/xmark.svg?raw';
 
-const icons = { clone, trash, xmark } as const;
+const icons = {
+  circleQuestion,
+  clone,
+  ellipsisVertical,
+  fileExport,
+  fileImport,
+  trash,
+  xmark,
+} as const;
 
 defineProps<{ name: keyof typeof icons }>();
 </script>